move fail_on to its own module
